Consolidate cluster Pydantic models into rpc.models; widen send opts
Move the Pydantic ClusterUrls/Endpoint into solana.rpc.models and delete
solana.utils.models, so all opts/info/cluster models share one home.
cluster.py imports them aliased (the bare names collide with the deprecated
NamedTuples it still defines for back-compat) and builds ENDPOINT from them;
the deprecation messages now point at solana.rpc.models.

Also widen send_raw_transaction/send_transaction opts to
Optional[Union[types.TxOpts, TxOptsModel]] so callers can still pass the
deprecated NamedTuple TxOpts.
