+++ src/solana/utils/cluster.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 from typing import Literal, NamedTuple
 from typing_extensions import deprecated
 
-from solana.utils import models as cluster_models
+from solana.rpc.models import ClusterUrls as ClusterUrlsModel, Endpoint as EndpointModel
 
 
-@deprecated("ClusterUrls is deprecated; use solana.utils.models instead.")
+@deprecated("ClusterUrls is deprecated; use solana.rpc.models instead.")
 class ClusterUrls(NamedTuple):
     """A collection of urls for each cluster."""
 
@@ -17,7 +17,7 @@ class ClusterUrls(NamedTuple):
     mainnet_beta: str
 
 
-@deprecated("Endpoint is deprecated; use solana.utils.models instead.")
+@deprecated("Endpoint is deprecated; use solana.rpc.models instead.")
 class Endpoint(NamedTuple):
     """Container for http and https cluster urls."""
 
@@ -25,13 +25,13 @@ class Endpoint(NamedTuple):
     https: ClusterUrls
 
 
-ENDPOINT = cluster_models.Endpoint(
-    http=cluster_models.ClusterUrls(
+ENDPOINT = EndpointModel(
+    http=ClusterUrlsModel(
         devnet="http://api.devnet.solana.com",
         testnet="http://api.testnet.solana.com",
         mainnet_beta="http://api.mainnet-beta.solana.com/",
     ),
-    https=cluster_models.ClusterUrls(
+    https=ClusterUrlsModel(
         devnet="https://api.devnet.solana.com",
         testnet="https://api.testnet.solana.com",
         mainnet_beta="https://api.mainnet-beta.solana.com/",
